src/test/java/com/down2thewire/RouteIdentifierTest.java
package com.down2thewire;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;
import java.util.Arrays;
import java.util.LinkedList;

class RouteIdentifierTest {

    private RouteIdentifier routeIdentifier;
    private BranchGeoModel geoModel;
    private UserRouteRequest routeRequest;


    @BeforeEach
    void setUp() {
        // Initialize the geoModel and routeRequest with mock data or mocks
        geoModel = new BranchGeoModel();
        routeRequest = new UserRouteRequest();


        // Initialize RouteIdentifier with the mock data
        routeIdentifier = new RouteIdentifier(geoModel, routeRequest);
    }

    @Test
    void testAddOperationAndListRetrieval() {
        DijkstraGraph graph = new DijkstraGraph(geoModel, routeRequest, "mode", "metric");
        routeIdentifier.addGraph(graph);
        LinkedList<DijkstraGraph> graphList = routeIdentifier.getGraphList();

        assertTrue(graphList.contains(graph));
    }

    @Test
    void testGetBestRoutePrimaryModeOnly() {
        // Assuming mock data where primary mode is sufficient

        // Call the method under test
        LinearRoute result = routeIdentifier.getBestRoute("car", "walking", "fastest", 10000);

        // Perform assertions to verify the result
        assertNotNull(result);
        // Add more detailed checks based on the expected behavior of the method
    }

    @Test
    void testGetBestRouteWithSecondaryMode() {
        // Assuming mock data where secondary mode is necessary

        // Call the method under test
        LinearRoute result = routeIdentifier.getBestRoute("car", "bus", "cheapest", 5000);

        // Perform assertions to verify the result
        assertNotNull(result);
        // Add more detailed checks based on the expected behavior of the method
    }

    @Test
    void testGetBestRouteNoPossiblePath() {
        // Assuming mock data where no path is possible

        // Call the method under test
        LinearRoute result = routeIdentifier.getBestRoute("car", "bus", "cheapest", 500);

        // Perform assertions to verify the result
        assertNull(result); // Assuming method returns null or throws an exception when no path is possible
        // If an exception is expected, use assertThrows
    }

//    @Test
//    void testGetAlternativeVertex() {
//        // Assuming mock data
//
//        // Call the method under test
//        LinearRoute result = routeIdentifier.getAlternativeVertex("car", "bus", "cheapest", "VertexX");
//
//        // Perform assertions to verify the result
//        assertNotNull(result);
//        // Add more detailed checks based on the expected behavior of the method
//    }

    // Additional tests for combineRoutes, calculateTotalDistance, etc. can be added here

    // You could also add more tests for edge cases, invalid inputs, etc.

}
